requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
torch==2.0.1
torchvision==0.15.2
pillow==10.3.0
numpy==1.26.4
opencv-python==4.9.0.80
python-multipart
gdown
matplotlib
